refactor: move mapConcrete into typeHelpers
