+++ src/typeEvaluator/typeEvaluate.ts
@@ -47,7 +47,7 @@ import type {
   UnionTypeNode,
   UnknownTypeNode,
 } from './types'
-import {nullUnion} from './typeHelpers'
+import {mapConcrete, nullUnion} from './typeHelpers'
 
 const $trace = debug('typeEvaluator:evaluate:trace')
 $trace.log = console.log.bind(console) // eslint-disable-line no-console
@@ -1084,54 +1084,6 @@ function mapUnion(node: TypeNode, mapper: (node: TypeNode) => TypeNode): TypeNod
     })
   }
   return mapper(node)
-}
-
-type ConcreteTypeNode =
-  | BooleanTypeNode
-  | NullTypeNode
-  | NumberTypeNode
-  | StringTypeNode
-  | ArrayTypeNode
-  | ObjectTypeNode
-
-/**
- * mapConcrete extracts a _concrete type_ from a type node, applies the mapping
- * function to it and returns. Most notably, this will work through unions
- * (applying the mapping function for each variant) and inline (resolving the
- * reference).
- *
- * An `unknown` input type causes it to return `unknown` as well.
- *
- * After encountering unions the resulting types gets passed into `mergeUnions`.
- * By default this will just union them together again.
- */
-function mapConcrete(
-  node: TypeNode,
-  scope: Scope,
-  mapper: (node: ConcreteTypeNode) => TypeNode,
-  mergeUnions: (nodes: TypeNode[]) => TypeNode = (nodes) =>
-    optimizeUnions({type: 'union', of: nodes}),
-): TypeNode {
-  switch (node.type) {
-    case 'boolean':
-    case 'array':
-    case 'null':
-    case 'object':
-    case 'string':
-    case 'number':
-      return mapper(node)
-    case 'unknown':
-      return node
-    case 'union':
-      return mergeUnions(node.of.map((inner) => mapConcrete(inner, scope, mapper), mergeUnions))
-    case 'inline': {
-      const resolvedInline = scope.context.lookupTypeDeclaration(node)
-      return mapConcrete(resolvedInline, scope, mapper, mergeUnions)
-    }
-    default:
-      // @ts-expect-error
-      throw new Error(`Unknown type: ${node.type}`)
-  }
 }
 
 function mapArray(
