+++ src/typeEvaluator/typeHelpers.ts
@@ -1,4 +1,16 @@
-import type {ObjectAttribute, ObjectTypeNode, TypeNode, UnionTypeNode} from './types'
+import {optimizeUnions} from './optimizations'
+import type {Scope} from './scope'
+import type {
+  ArrayTypeNode,
+  BooleanTypeNode,
+  NullTypeNode,
+  NumberTypeNode,
+  ObjectAttribute,
+  ObjectTypeNode,
+  StringTypeNode,
+  TypeNode,
+  UnionTypeNode,
+} from './types'
 
 /**
  * createReferenceTypeNode creates a ObjectTypeNode representing a reference type
@@ -84,5 +96,53 @@ export function resolveUnionType(
     default: {
       return resolver(typeNode)
     }
+  }
+}
+
+type ConcreteTypeNode =
+  | BooleanTypeNode
+  | NullTypeNode
+  | NumberTypeNode
+  | StringTypeNode
+  | ArrayTypeNode
+  | ObjectTypeNode
+
+/**
+ * mapConcrete extracts a _concrete type_ from a type node, applies the mapping
+ * function to it and returns. Most notably, this will work through unions
+ * (applying the mapping function for each variant) and inline (resolving the
+ * reference).
+ *
+ * An `unknown` input type causes it to return `unknown` as well.
+ *
+ * After encountering unions the resulting types gets passed into `mergeUnions`.
+ * By default this will just union them together again.
+ */
+export function mapConcrete(
+  node: TypeNode,
+  scope: Scope,
+  mapper: (node: ConcreteTypeNode) => TypeNode,
+  mergeUnions: (nodes: TypeNode[]) => TypeNode = (nodes) =>
+    optimizeUnions({type: 'union', of: nodes}),
+): TypeNode {
+  switch (node.type) {
+    case 'boolean':
+    case 'array':
+    case 'null':
+    case 'object':
+    case 'string':
+    case 'number':
+      return mapper(node)
+    case 'unknown':
+      return node
+    case 'union':
+      return mergeUnions(node.of.map((inner) => mapConcrete(inner, scope, mapper), mergeUnions))
+    case 'inline': {
+      const resolvedInline = scope.context.lookupTypeDeclaration(node)
+      return mapConcrete(resolvedInline, scope, mapper, mergeUnions)
+    }
+    default:
+      // @ts-expect-error - we should handle each type
+      throw new Error(`Unknown type: ${node.type}`)
   }
 }
